GH-3336: Change MongoDb Store sequence to long (#3385)
* GH-3336: Change MongoDb Store sequence to long

Fixes https://github.com/spring-projects/spring-integration/issues/3336

Turns out there are some scenarios where too many messages
are transferred through the message store, so `int` for
sequence is not enough as a type

* Change sequence to `long` to widen a sequence lifespan

* * Change MongoDb store to deal with `Number.longValue()`
instead of casting which doesn't work from `Integer` to `Long`.
This way we can keep an old sequence document with an `int`
type for value
* Documents with new `long` type for their sequence field are OK.
The `NumberToNumberConverter` has an effect converting `int` to `long`
properly.

+++ spring-integration-mongodb/src/main/java/org/springframework/integration/mongodb/store/MessageDocument.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2014-2019 the original author or authors.
+ * Copyright 2014-2020 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -61,7 +61,7 @@ public class MessageDocument {
 
 	private Integer lastReleasedSequence = 0;
 
-	private int sequence;
+	private long sequence;
 
 	public MessageDocument(Message<?> message) {
 		this(message, message.getHeaders().getId());
@@ -139,7 +139,7 @@ public class MessageDocument {
 		this.lastReleasedSequence = lastReleasedSequence;
 	}
 
-	public void setSequence(int sequence) {
+	public void setSequence(long sequence) {
 		this.sequence = sequence;
 	}
 
@@ -151,7 +151,7 @@ public class MessageDocument {
 		return this.groupId;
 	}
 
-	public int getSequence() {
+	public long getSequence() {
 		return this.sequence;
 	}
